AP-6093: Restructure system tests
