+++ spec/system/application_task_list/client_and_case_details_section/check_provider_answers_spec.rb
@@ -1,0 +1,114 @@
+require "system_helper"
+RSpec.describe "Client and case details section - Check your answers", :vcr do
+  feature "View check your answers" do
+    before do
+      login_as_a_provider
+
+      Setting.setting.update!(linked_applications: true)
+    end
+
+    scenario "I can complete the task list's Check your answers item" do
+      fill_in_client_and_case_details_until_step(:has_national_insurance_number)
+
+      expect(page).to have_css("h1", text: "Does your client have a National Insurance number?")
+
+      @legal_aid_application ||= LegalAidApplication.find(id_from_current_page_url)
+      visit providers_legal_aid_application_task_list_path(@legal_aid_application)
+
+      expect_section_with_task_list_items("Client and case details") do
+        [
+          { name: "Client details", link_enabled: true, status: "In progress" },
+          { name: "Link to another application", link_enabled: false, status: "Cannot start yet" },
+          { name: "Check your answers", link_enabled: false, status: "Not ready yet" },
+        ]
+      end
+
+      click_link "Client details"
+      click_on "Save and continue"
+
+      govuk_choose("Yes")
+      fill_in("Enter National Insurance number", with: "CB987654A")
+      click_on "Save and continue"
+
+      expect(page).to have_css("h1", text: "Has your client applied for civil legal aid before?")
+      govuk_choose("No")
+      click_on "Save and continue"
+
+      govuk_choose("My client's UK home address")
+      click_on "Save and continue"
+
+      fill_in("Postcode", with: "SW1H 9EA")
+      click_on "Find address"
+      govuk_choose "Transport For London, 98 Petty France, London, SW1H 9EA"
+      click_on "Use this address"
+
+      expect(page).to have_css("h1", text: "Do you want to link this application with another one?")
+      govuk_choose("No")
+      click_on "Save and continue"
+
+      visit providers_legal_aid_application_task_list_path(@legal_aid_application)
+
+      expect_section_with_task_list_items("Client and case details") do
+        [
+          { name: "Client details", link_enabled: false, status: "Completed" },
+          { name: "Link to another application", link_enabled: false, status: "Completed" },
+          { name: "Check your answers", link_enabled: true, status: "Not started" },
+        ]
+      end
+
+      click_on "Check your answers"
+      click_on "Save and come back later"
+
+      visit providers_legal_aid_application_task_list_path(@legal_aid_application)
+
+      expect_section_with_task_list_items("Client and case details") do
+        [
+          { name: "Client details", link_enabled: false, status: "Completed" },
+          { name: "Link to another application", link_enabled: false, status: "Completed" },
+          { name: "Check your answers", link_enabled: true, status: "In progress" },
+        ]
+      end
+
+      click_on "Check your answers"
+      click_on "Save and continue"
+
+      visit providers_legal_aid_application_task_list_path(@legal_aid_application)
+
+      expect_section_with_task_list_items("Client and case details") do
+        [
+          { name: "Client details", link_enabled: false, status: "Completed" },
+          # { name: "Link to another application", link_enabled: false, status: "Completed" }, # TODO: fix as part of 6091
+          { name: "Check your answers", link_enabled: false, status: "Completed" },
+        ]
+      end
+
+      # Note that this emulates the eventual ability of a user to go back to a completed task list item which
+      # currently is not enabled as "Completed" items are not navigable from the "restrictive" task list.
+      visit providers_legal_aid_application_applicant_details_path(@legal_aid_application)
+      click_on "Save and continue"
+
+      visit providers_legal_aid_application_task_list_path(@legal_aid_application)
+
+      expect_section_with_task_list_items("Client and case details") do
+        [
+          { name: "Client details", link_enabled: false, status: "Completed" },
+          # { name: "Link to another application", link_enabled: false, status: "Completed" }, # TODO: fix as part of 6091
+          { name: "Check your answers", link_enabled: true, status: "Review" },
+        ]
+      end
+
+      within(section_list_for("Client and case details")) { click_link "Check your answers" }
+      click_on "Save and continue"
+
+      visit providers_legal_aid_application_task_list_path(@legal_aid_application)
+
+      expect_section_with_task_list_items("Client and case details") do
+        [
+          { name: "Client details", link_enabled: false, status: "Completed" },
+          # { name: "Link to another application", link_enabled: false, status: "Completed" }, # TODO: fix as part of 6091
+          { name: "Check your answers", link_enabled: false, status: "Completed" },
+        ]
+      end
+    end
+  end
+end
